Refactor sharedInstance; Add Continuity support
Refactored ApplicationSupervisor.sharedInstance to return the shared application’s app delegate. This allows the ApplicationSupervisor to be instantiated directly for unit testing, and also couples the API to UIApplication, as ApplicationSupervisor is meant to be the app delegate.

Added support for Continuity by extending the PluggableFeature protocol and dispatching to plugins that implement the API.

Added a unit test to test the dispatch mechanism on ApplicationSupervisor. Added a testPluginAPI to capture test results for later inspection by a unit test.

Added the unit test target to the ELMaestro scheme’s test phase.